Split models in their own file
