+++ hub/admin.py
@@ -9,7 +9,7 @@ from django.contrib.auth.forms import (
     UserCreationForm as DjangoUserCreationForm,
     UserChangeForm as DjangoUserChangeForm,
 )
-from django.contrib.auth.models import User
+from django.contrib.auth.models import User, Group
 from django.core.exceptions import ValidationError
 from django.db.models import Count, Sum
 from django.forms import CharField
@@ -344,5 +344,6 @@ admin.site.register(ExtraUserDetail, ExtraUserDetailAdmin)
 admin.site.register(SitewideMessage)
 admin.site.register(ConfigurationFile)
 admin.site.register(PerUserSetting)
+admin.site.unregister(Group)
 admin.site.unregister(User)
 admin.site.register(User, ExtendedUserAdmin)
