+++ hub/models/__init__.py
@@ -1,0 +1,8 @@
+from .configuration_file import (
+    _configuration_file_upload_to,  # needed to register model below
+    ConfigurationFile,
+    ConfigurationFileSlug,
+)
+from .extra_user_detail import ExtraUserDetail
+from .sitewide_message import SitewideMessage
+from .per_user_setting import PerUserSetting
