+++ hub/models/sitewide_message.py
@@ -1,0 +1,10 @@
+from django.db import models
+from markitup.fields import MarkupField
+
+
+class SitewideMessage(models.Model):
+    slug = models.CharField(max_length=50)
+    body = MarkupField()
+
+    def __str__(self):
+        return self.slug
